|FEATURE| finished the Branch and refactored the other entities

+++ backend/src/main/java/com/ksu/inventory_management_system_backend/entities/supplier/SupplierServiceImpl.java
@@ -1,9 +1,11 @@
 package com.ksu.inventory_management_system_backend.entities.supplier;
 
-
 import com.ksu.inventory_management_system_backend.entities.supplier.dto.CreateSupplierDto;
+import com.ksu.inventory_management_system_backend.exceptions.ReferentialIntegrityViolationException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -12,28 +14,41 @@ public class SupplierServiceImpl implements SupplierService{
     private final JpaSupplierRepository jpaSupplierRepository;
 
     @Override
-    public Supplier createSupplier(CreateSupplierDto createSupplierDto) {
-        return jpaSupplierRepository.save(Supplier.from(createSupplierDto));
+    public void createSupplier(CreateSupplierDto createSupplierDto) {
+        Supplier supplier = Supplier.from(createSupplierDto);
+        jpaSupplierRepository.save(supplier);
     }
 
     @Override
-    public Supplier updateSupplier(long id, CreateSupplierDto createSupplierDto) {
-        return jpaSupplierRepository.save(Supplier.from(createSupplierDto));
-    }
-
-    @Override
-    public Supplier getSupplier(long id) {
-        return jpaSupplierRepository.findById(id);
+    public void updateSupplier(long id, CreateSupplierDto createSupplierDto) {
+        Optional<Supplier> optionalSupplier = jpaSupplierRepository.findById(id);
+        if (optionalSupplier.isPresent()) {
+            Supplier supplier = optionalSupplier.get();
+            supplier.setName(createSupplierDto.name());
+            supplier.setContactNumber(createSupplierDto.contactNumber());
+            supplier.setAddress(createSupplierDto.address());
+            jpaSupplierRepository.save(supplier);
+        } else {
+            throw new ReferentialIntegrityViolationException("Supplier with id " + id + " does not exist");
+        }
     }
 
     @Override
     public void deleteSupplier(long id) {
-        jpaSupplierRepository.deleteById(id);
+        if (jpaSupplierRepository.existsById(id)) {
+            jpaSupplierRepository.deleteById(id);
+        } else {
+            throw new ReferentialIntegrityViolationException("Supplier with id " + id + " does not exist");
+        }
     }
 
     @Override
     public Supplier getSupplierByName(String name) {
-        return jpaSupplierRepository.findByName(name);
+        return jpaSupplierRepository.findByName(name).orElseThrow(() -> new ReferentialIntegrityViolationException("Supplier with name " + name + " does not exist"));
     }
 
+    @Override
+    public Supplier getSupplier(long id) {
+        return jpaSupplierRepository.findById(id).orElseThrow(() -> new ReferentialIntegrityViolationException("Supplier with id " + id + " does not exist"));
+    }
 }
